remove code for test from jwt-auth.lua
